Exercise 4.21

+++ controllers/blogs.js
@@ -48,8 +48,26 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  const user = await User.findById(decodedToken.id);
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    user.blogs.filter(b => b.toString() !== blog._id.toString());
+    await user.save();
+    return response.status(204).end();
+  } else {
+    return response.status(401).json({
+      error: 'unauthorized attempt to delete'
+    });
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
